pyproject.toml
[project]
name = "digpipe"
version = "0.1.0"
description = ""
authors = [
    {name = "DJ Stomp",email = "85457381+DJStompZone@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
]

[tool.poetry]
packages = [{include = "digpipe"}]


[tool.poetry.group.dev.dependencies]
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
